机器人初步测试代码

+++ baseLibrary/enginesdk/src/main/java-cb/media/ushow/score/ScoreProcessor.java
@@ -33,7 +33,9 @@ public class ScoreProcessor {
     }
 
     public void destroy() {
-        mScoreProcessorService.destroy();
-        mScoreProcessorService = null;
+        if(mScoreProcessorService != null){
+            mScoreProcessorService.destroy();
+            mScoreProcessorService = null;
+        }
     }
 }
